+++ client/cypress.config.js
@@ -2,9 +2,12 @@ const { defineConfig } = require("cypress");
 
 module.exports = defineConfig({
   e2e: {
-    baseUrl:'http://localhost:3000',
-    setupNodeEvents(on, config) {
-      // implement node event listeners here
+    baseUrl: "http://localhost:3000",
+    env: {
+      credentials: {
+        username: "gary.cole@example.com",
+        password: "pAssw0rd",
+      },
     },
   },
 });